chore(backend): Move /tokens/jwt to top-level and replace jwt.ts
