--- packages/backend/src/jwt.ts
@@ -1,5 +1,0 @@
-export { hasValidSignature, decodeJwt, verifyJwt } from './tokens/jwt/verifyJwt';
-export { signJwt } from './tokens/jwt/signJwt';
-
-export type { VerifyJwtOptions } from './tokens/jwt/verifyJwt';
-export type { SignJwtOptions } from './tokens/jwt/signJwt';
