+++ packages/backend/src/tokens/__tests__/verify.test.ts
@@ -1,9 +1,9 @@
 import type QUnit from 'qunit';
 import sinon from 'sinon';
 
+import { mockJwks, mockJwt, mockJwtPayload } from '../../fixtures';
 import runtime from '../../runtime';
 import { jsonOk } from '../../util/testUtils';
-import { mockJwks, mockJwt, mockJwtPayload } from '../fixtures';
 import { verifyToken } from '../verify';
 
 export default (QUnit: QUnit) => {
